+++ src/imgui/backends/imgui_d3d9.cpp
@@ -685,10 +685,10 @@ ImGui_ImplDX9_NewFrame (void)
 
   io.DeltaTime =
     std::min ( 1.0f,
-    std::max ( 0.0f, static_cast <float> (
-                    (static_cast <long double> (                       current_time) -
-                     static_cast <long double> (std::exchange (g_Time, current_time))) /
-                     static_cast <long double> (               g_TicksPerSecond      ) ) )
+    std::max ( 0.0f, static_cast <float>  (
+                    (static_cast <double> (                       current_time) -
+                     static_cast <double> (std::exchange (g_Time, current_time))) /
+                     static_cast <double> (               g_TicksPerSecond      ) ) )
     );
 
   // Read keyboard modifiers inputs
